requirements.txt
scikit-learn
numpy
openai
scipy
python-dotenv
dash
dash-bootstrap-components
# Add other dependencies here, one per line
# For example:
# pandas
# flask
